ipc: make each ipc call use its own connection to socket
I am not sure why, but reusing the same connection would make the child
process unresponsive and have very high cpu usage. I suspect that it was
due to some limitation in the way it was coded, as testing the ipc
connection does not have any issue outside the program.