ipc: read response from mpv when issues commands
It appears that the issue I had in the previous commit was that I was
not reading anything from the stream. This could cause the stream to
overflow when to many writes had been done, although is my take with no
confirmation. After reading from the stream, I am not able to reproduce
the issue. And it is more elegant to only have one shared stream for all
ipc calls :)