app,ipc: get playback time from mpv
This uses the mpv command to get the playback time, since I noticed that
there were deviations when using the application's own playback time and
too many seek commands. This method make use of the seek events to
update the playback time.

Also noteworthy is that I needed to add a delay to parse the response
from the get_playback command, or else there would benot enough time for
mpv to write the response in the stream.