app, server: back to caching all albums
I realized that the sibsonic API does not allow to search for albumList
and sort by most played, only alphabetical order. The sorting has to be
done locally, and in order to do so, I need to know all genres of all
albums. Therefore, I need to have the complete information of all albums
before starting the app.

This has some implications:
  1. Very slow startup, therefore next thing will be to persist to disk
     the database
  2. The need to maintain consistency between local database and
     server's