app: allow for offline mode starting and error handling
Specifically, the error handling is regarding the initializing of the
ipc stream. The startup was being too fast for mpv to accept the ipc
connection. This was due to skipping the server connection calls in
offline mode, which were giving enough time for mpv to get up. Now,
there is a retry system for the ipc stream with a 30ms delay.